ccl(provenance): one module, one word for where a node came from
`lineage` and `provenance` were two names for one concept, and the split file
followed the split word. `src/ccl/lineage.rs` folds into `src/ccl/provenance.rs`,
and the word is gone from prose, identifiers and environment variables alike.

**The stated reason for the split does not hold.** `provenance.rs` claimed to own
"the two identity primitives that lineage is parameterized over", but `Pass` is
not an identity primitive — it is a row's `via` column, which the same doc comment
said outright. `NodeId` was the only genuine resident. Nor was either module's
doc readable alone: twenty of `provenance.rs`'s twenty-one module-doc lines
explained lineage before disclaiming it, and the two were imported by nearly the
same set of modules (15 and 18, overlapping in 8), so there was no
wide-primitive/narrow-machinery boundary to keep.

**The word carried no distinction either.** `provenance.md` said "lineage"
twenty-two times and "provenance" three, one of those being its own title. Where
the word did mean something narrower than the whole subsystem it meant the
**ancestry** axis — which `EdgeLabels::ANCESTRY` already names, making "lineage" a
third name for it. Those uses now say `ancestry`, and the axis vocabulary is
`ancestry` / `blame` / `spans` throughout.

## What moved

`LineageTable` is `ProvenanceTable`, `LineageMap` is `ProvenanceMap`,
`LineageAudit` is `ProvenanceAudit`, `CompiledProgram::lineage_table` is
`provenance_table`, and `lineage_capture_enabled` and its three sibling switch
readers follow. The four environment variables are `CAMBRA_PROVENANCE`,
`CAMBRA_PROVENANCE_AUDIT`, `CAMBRA_PROVENANCE_GATE` and
`CAMBRA_PROVENANCE_PREDICATES`; `ci.yml` names the third.

The merged file opens with one module doc rather than two mutually-deferring
ones, and a banner marks where identity ends and the table begins. `Pass`'s doc
regains the pipeline order it used to delegate to the module doc, and drops the
coined "WHO axis", which named nothing and had no sibling axes.

`provenance.md` stops citing the file in three sibling headings — one file backs
all three now — and says once, up front, which file holds what. `#the-node-table`
and `#the-provenance-model` are the anchors; `doc_refs` checks both, plus the two
code citations that name those headings.

The vault reference in `tests/predicate_sharing.rs` had a second spelling of the
note `type-inference.md` cites, carrying a cross-doc section number with it. Both
now read the same way.

`ci.sh` green in all four clippy configurations plus doc, doc_refs and shellcheck,
with the suite passing under `CAMBRA_PROVENANCE_GATE=1`.
