+++ src/ccl/mod.rs
@@ -13,7 +13,6 @@ pub mod infer;
 pub mod inline;
 pub mod lambda_elim;
 pub mod letrec;
-pub mod lineage;
 pub mod lower;
 pub mod mut_elim;
 pub mod names;
